scripts/requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
packaging>=24.0
